refactor modules for readability
