+++ backend/app/strands_integration/handlers/__init__.py
@@ -1,0 +1,11 @@
+"""
+Handlers module for Strands integration.
+"""
+from .callback_handler import CallbackHandler, create_callback_handler
+from .tool_result_capture import ToolResultCapture
+
+__all__ = [
+    "CallbackHandler",
+    "create_callback_handler",
+    "ToolResultCapture",
+]
