+++ backend/app/strands_integration/processors/__init__.py
@@ -1,0 +1,19 @@
+"""
+Processors module for Strands integration.
+"""
+from .cost_calculator import calculate_conversation_cost
+from .document_extractor import (
+    build_thinking_log_from_tool_capture,
+    extract_reasoning_from_message,
+    extract_related_documents_from_tool_capture,
+)
+from .result_processor import create_on_stop_input, post_process_strands_result
+
+__all__ = [
+    "calculate_conversation_cost",
+    "extract_related_documents_from_tool_capture",
+    "build_thinking_log_from_tool_capture",
+    "extract_reasoning_from_message",
+    "create_on_stop_input",
+    "post_process_strands_result",
+]
